Rename MiniDay to MiniDaySwitch
