+++ src/components/UI/MiniDaySwitch.tsx
